+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Check for exceeding of maxIssuanceAmount on create `CREATE_SALE` request
 * Remove checking for not empty signers on signature validation
 * Return error code on review pre issuance if max issuance amount exceeded
+* ensureValid method of asset frame, (add availableForIssuance to sum which will be compared with maxIssuance)
 
 ## 3.0.0-x.8