+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+* Exception on checkSaleState with cause allowing pre issuance excess tokens after sale creation
+* Exception on check signature when there is no signers for account
+* Exception on review pre issuance when available for issuance more than max issuance
+
 ## 3.0.0-x.8
 
 ### Fixed
@@ -20,6 +28,7 @@
 * Field with payload for txNO_ROLE_PERMISSION and opNO_ROLE_PERMISSION
 * Licensing
 * License, Stamp operations
+
 ### Fixed
 
 * Do not add to cache if load entry without delta (add delta to putCacheEntry method)